requirements-full.txt
# 由 requirements.txt 在干净 venv 中 pip install 后 pip freeze 生成（含全部传递依赖，版本锁定）。
# 安装：pip install -r requirements-full.txt
#
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
crcmod==1.7
cryptography==46.0.6
Deprecated==1.3.1
ecdsa==0.19.2
exceptiongroup==1.3.1
fastapi==0.115.0
frozenlist==1.8.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.1
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.2
idna==3.11
json_repair==0.44.1
multidict==6.7.1
packaging==26.0
passlib==1.7.4
pillow==11.3.0
playwright==1.58.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.9.0
pydantic-settings==2.6.1
pydantic_core==2.23.2
pyee==13.0.1
PyJWT==2.12.1
pyparsing==3.3.2
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.19
pytz==2026.1.post1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.36
starlette==0.38.6
tos==2.9.0
twilio==9.10.4
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
wrapt==1.16.0
yarl==1.22.0
